Remove unused files
- Removed unused DeviceRowItem
- Removed General in favour of index.js in utils
- Remove unused styles in DeviceList
