--- src/components/devices/DeviceRowItem.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import DeviceRow from './DeviceRow';
-
-const DeviceRowItem: () => React$Node = props => {
-  return <DeviceRow {...props} />;
-};
-
-export default DeviceRowItem;
