+++ src/util/index.js
@@ -1,3 +1,10 @@
+const LastUploadTime = lastUploaded => {
+  const delta = new Date() - new Date(lastUploaded);
+  const days = Math.floor(delta / (1000 * 60 * 60 * 24));
+  const hours = new Date(delta).getHours();
+  return days > 0 ? `${days} days ago` : `${hours} hours ago`;
+};
+
 const FormatBytes = (bytes, decimals = 0) => {
   const k = 1024;
   const dm = decimals < 0 ? 0 : decimals;
@@ -7,4 +14,4 @@ const FormatBytes = (bytes, decimals = 0) => {
   return `${size} ${sizes[i]}`;
 };
 
-export {FormatBytes};
+export {LastUploadTime, FormatBytes};
